add identity settings

+++ Src/App/zabotalaboratory/zabotalaboratory.Auth.Services/Identities/IdentitiesService.cs
@@ -3,10 +3,10 @@ using AutoMapper;
 using zabotalaboratory.Auth.Database.Repository.Identities;
 using System.Collections.Generic;
 using zabotalaboratory.Auth.Datamodel.Identities;
-using zabotalaboratory.Auth.Forms.Login;
 using zabotalaboratory.Common.Result;
-using zabotalaboratory.Auth.Database.Entities;
 using zabotalaboratory.Auth.Datamodel.Roles;
+using zabotalaboratory.Common.Result.ErrorCodes;
+using zabotalaboratory.Auth.Forms.Identity;
 
 namespace zabotalaboratory.Auth.Services.Identities
 {
@@ -29,14 +29,24 @@ namespace zabotalaboratory.Auth.Services.Identities
             return new ZabotaResult<IEnumerable<ZabotaIdentity>>(mappedIdentities);
         }
 
-        public async Task<ZabotaIdentity> GetIdentityByTokenId(int id)
+        public async Task<ZabotaResult<ZabotaIdentity>> GetIdentityById(int id)
+        {
+            var model = await _identitiesRepository.IdentityById(id);
+            if (model == null)
+                return ZabotaErrorCodes.CannotFindIdentityById;
+
+            var mappedModel = _mapper.Map<Database.Entities.Identities, ZabotaIdentity>(model);
+            return new ZabotaResult<ZabotaIdentity>(mappedModel);
+        }
+
+        public async Task<ZabotaResult<ZabotaIdentity>> GetIdentityByTokenId(int id)
         {
             var model = await _identitiesRepository.IdentityByTokenId(id);
             if (model == null)
-                return null;
+                return ZabotaErrorCodes.CannotFindIdentityByTokenId;
 
             var mappedModel = _mapper.Map<Database.Entities.Identities, ZabotaIdentity>(model);
-            return mappedModel;
+            return new ZabotaResult<ZabotaIdentity>(mappedModel);
         }
 
         public async Task<ZabotaResult<IEnumerable<ZabotaRoles>>> GetRoles() 
@@ -55,21 +65,35 @@ namespace zabotalaboratory.Auth.Services.Identities
             return new ZabotaResult<IEnumerable<ZabotaSubRoles>>(mappedModel);
         }
 
-        public async Task<bool> AddIdentity(LoginForm form)
+        public async Task<ZabotaResult<bool>> AddIdentity(AddIdentityForm form)
         {
             var model = await _identitiesRepository.IdentityByLogin(form.Login);
             if (model != null)
-                return false;
+                return ZabotaErrorCodes.IdentityExists;
 
             await _identitiesRepository.Add(form);
-            return true;
+            return new ZabotaResult<bool>(true);
         }
 
-        public async Task<bool> DeleteIdentity(int id, int actorId)
+        public async Task<ZabotaResult<bool>> UpdateIdentity(UpdateIdentityForm form)
         {
-            var result = await _identitiesRepository.Delete(id, actorId);
+            var result = await _identitiesRepository.Update(form);
+            if (result != true)
+                return ZabotaErrorCodes.CannotFindIdentityById;
 
-            return result;
+            return new ZabotaResult<bool>(true);
+        }
+
+        public async Task<ZabotaResult<bool>> DeleteIdentity(int id, int actorId)
+        {
+            if (id == actorId)
+                return ZabotaErrorCodes.CannotDeleteCurrentUser;
+
+            var result = await _identitiesRepository.Delete(id, actorId);
+            if (result != true)
+                return ZabotaErrorCodes.CannotFindIdentityById;
+
+            return new ZabotaResult<bool>(true);
         }
     }
 }
